==> 02 Text Preprocessing.ipynb <==
# Text preprocessing

## Necessary downloads and library imports

from google.colab import drive
drive.mount('/content/drive')

project_path = '/content/drive/My Drive/Colab Notebooks/MATF_ML_project/'

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

import pickle

from wordcloud import WordCloud

from bs4 import BeautifulSoup
import nltk
from nltk.corpus import wordnet
from nltk.corpus import stopwords
import regex as re
import string

nltk.download('punkt')
nltk.download('stopwords')
!pip install lxml

## Load data

We will turn off the detection of missing value markers due to the presence of "null" tags that are otherwise recognised as NA values. There are no true NA values in the data.

In order to avoid computation constraints with training and evaluation we will use a small subset of the data. Only 10,000 instances.

data = pd.read_csv(project_path + 'data/Train.csv',
                   na_filter=False, nrows=10000)

## Duplicates in data

data_without_duplicates = data.drop_duplicates(['Title', 'Body', 'Tags'])

print("Shape with duplicates\t", data.shape)
print("Shape without duplicates", data_without_duplicates.shape)

data = data_without_duplicates

### Drop rare tags

In order to avoid computation constraints with training and evaluation of the data, we will limit our tag prediction to the top 100 most frequent tags.

N_tags = 100

tags_per_question = list(map(lambda tags: tags.split(' '), data["Tags"]))
all_tags = [item for sublist in tags_per_question for item in sublist]
all_tags = np.array(all_tags)
print(len(np.unique(all_tags)))
print("Tags:", all_tags)

unique, counts = np.unique(all_tags, return_counts=True)
tag_counts = dict(zip(unique, counts))

list(tag_counts.items())[:10]

tag_counts_sorted = dict(sorted(tag_counts.items(), key=lambda item: item[1], reverse=True))

print('Total number of tags is', len(tag_counts_sorted))

print("Most frequent tags:")
print(list(tag_counts_sorted.keys())[:20])

print("Rarest tags:")
print(list(tag_counts_sorted.keys())[-20:])

tags_most_frequent = list(tag_counts_sorted.keys())[:N_tags]

print("Top", N_tags, "tags:")
print(tags_most_frequent)

Drop rare tags from questions:

new_data_tags = []
for tags in data['Tags']:
    new_tags = [tag for tag in tags.split(" ") if tag in tags_most_frequent]
    if not new_tags:
        new_tags=None
    new_data_tags.append(new_tags)

data['Tags'] = new_data_tags

data.head()

data.shape

Drop instances that left without tags:

data = data.dropna()

data.shape

## Text preprocessing

test_str = "<body><p>There're some files: index.html, my.file.txt, file.c, file.ext-with-dash and I hope toktok-tokenizer will properly tokenize it.\n New line. This is new sentence. The most common tag is C#.... ++++++++++++++++++  ======================== R and C++ and C++11 and C aren't so frequent.<br> Some words with numbers: word1, word32, word-2.</p></body> There are some functions: main(), str(), f().<br>There are some urls: https://www.google.com https://www.google.com/dir/1/2/search.html?arg=0-a&arg1=1-b&arg3-c#hash https://google.us.edi?34535/534534?dfg=g&fg. Some numbers:  0123, 2021, 30912, 0000"

#### Removing HTML tags

We choose `lxml` believing it to be faster and more robust than the default one (```html.parser```).

def remove_html(s):
    return BeautifulSoup(s, 'lxml').get_text()

test_str = remove_html(test_str)
test_str

#### Cleaning text

Lowering text and transforming abbreviations.

file = open(project_path + "data/abbr.pkl", "rb")
abbr_dict = pickle.load(file)
file.close()

list(abbr_dict.items())[:10]

def clean_text(s, abbr_dict=None):
    s = s.lower()
    
    if abbr_dict is None:
      return s

    for patt, repl in abbr_dict.items():
        s = re.sub(patt, repl, s)
    return s

test_str = clean_text(test_str, abbr_dict)
test_str

#### Tokenization

We examined the following tokenizers on our data:
* ```nltk.tokenize.word_tokenize()```
* ```nltk.tokenize.WhitespaceTokenizer()```
* ```nltk.tokenize.ToktokTokenizer()```
* ```nltk.tokenize.TweetTokenizer()```
* ```nltk.tokenize.treebank.TreebankWordTokenizer()```

and TokTokTokenizer seemed most applicable.

word_tokenizer = nltk.tokenize.ToktokTokenizer()

def tokenize_text(s, word_tokenizer):
    sentences = nltk.tokenize.sent_tokenize(s)
    # suming lists, tokenize_sents returns list(list(str))
    return sum(word_tokenizer.tokenize_sents(sentences), [])

test_tokens = tokenize_text(test_str, word_tokenizer)
print(test_tokens)

#### Process URLs

We firstly considered extracting the URL into meaningful parts (host, path, query, ...) but it turned out that meaningless parts were dominant in number and that created noise in data.
So we decided to remove URLs at all.

def remove_urls(tokens):
    new_tokens = []
    url_pattern = r"^((http[s]?|ftp):\/)?\/?([^:\/\s]*)((\/\w+)*\/)([\w\-\.]+[^#?\s]+)(.*)?(#[\w\-]+)?$"
    for token in tokens:
        if not re.match(url_pattern, token):
            new_tokens.append(token)
    return new_tokens
    

test_tokens = remove_urls(test_tokens)
print(test_tokens)

#### Process extensions

File extensions are very important features because there are also file extensions as tags!

def extract_extensions(tokens):
    new_tokens = []
    for token in tokens:
        token = re.sub(r"([a-zA-Z0-9])\.([a-zA-Z0-9-])", r'\1 .\2', token)
        for new_token in token.split(" "):
            new_tokens.append(new_token)
    return new_tokens

test_tokens = extract_extensions(test_tokens)
print(test_tokens)

#### Punctuation

Removing punctuation but keeping:
* tokens like 'c#', 'c++', ...
* dots in file extensions (this significantly improved the metrics value!)

string.punctuation

punctuation_without_dash = '!"#$%&\'()*+,./:;<=>?@[\\]^_`{|}~'

def extension(word):
    if len(word) < 2:
        return False
    if word[0] != '.':
        return False
    for letter in word[1:]:
        if letter in punctuation_without_dash:
            return False
    return True

def remove_punctuation(tokens):
    new_tokens = []
    for token in tokens:
        if not extension(token):
            while len(token) != 0 and token[0] in string.punctuation:
                token = token[1:]
            while len(token) != 0 and token[-1] == '(':
                token = token[:-1]
        if len(token) != 0:
            new_tokens.append(token)
    return new_tokens

test_tokens = remove_punctuation(test_tokens)
print(test_tokens)

#### Stopwords

There are tags like 'design' so we think that tokens like 'why', 'how', 'what' are important (e.g. 'how' is maybe important for tag 'design') and we will not remove them. Also, we will not remove single-letter tokens (for example 'o').

stopwords_eng = stopwords.words('english')

print(stopwords_eng)

not_stopwords = ['what', 'which', 'who', 'about', 'where', 'why', 'how',
                  'no', 'not', 'on', 'off', 'o', 're', 'y']

my_stopwords = [word for word in stopwords_eng if word not in not_stopwords]
# my_stopwords

def remove_stopwords(tokens, stopwords):
    return [token for token in tokens if token not in stopwords]

test_tokens = remove_stopwords(test_tokens, my_stopwords)
print(test_tokens)

#### Numbers

To reduce the number of useless tokens we will reduce the numbers in text.
However, we will not completely eliminate them because they are important for questions from mathematical topics. We will replace number with one digit. 

In order not to lose the years from the text, which are also important, because they are often found in the names of software versions for example, we will save all the numbers between 1900 and 2100. 

To make sure of that, later in the notebook we will count occurrences of some years in the text.

NOTE: There is a lot of room for improvement in this approach. 

def process_numbers(tokens):
    new_tokens = []
    for token in tokens:
        while len(token) != 0 and token[0] == '0':
            token = token[1:]

        if token.isdigit() and len(token) == 4 and (int(token) > 1900 and int(token) < 2100): # token is PROBABLY a year
            new_tokens.append(token)
        elif token.isdigit():
            new_tokens.append(token[0])
        elif len(token) != 0:
            new_tokens.append(token)
    return new_tokens

test_tokens = process_numbers(test_tokens)
print(test_tokens)

#### Stemmer

We examined the following stemmers on our data:
* `nltk.stem.PorterStemmer(language='english')` - the most gentle one
* `nltk.stem.SnowballStemmer(language='english')` - improvement over porter
* `nltk.stem.lancaster.LancasterStemmer(language='english')` - the fastest but very agressive

We choose Snowball Stemmer.

We also examined lemmatizing but it was unacceptably slow.

stemmer = nltk.stem.SnowballStemmer(language='english')

def stem_tokens(tokens, stemmer):
    return [stemmer.stem(token) for token in tokens]

test_tokens = stem_tokens(test_tokens, stemmer)
print(test_tokens)

### Final preprocessing

def preprocess_text(text, tokenize=True, word_tokenizer=None, html=True, clean=True, abbr_dict = None, urls=True, extensions=True, 
                    punctuation=True, stopwords=True, stopword_list=None,
                    numbers=True, stem=True, stemmer=None):
    if html:
        text = remove_html(text)

    if abbr_dict is None:
        abbr_dict = {}

    if clean:
        text = clean_text(text, abbr_dict)
        
    if tokenize:
        if word_tokenizer is None:
            nltk.tokenize.ToktokTokenizer()
        tokens = tokenize_text(text, word_tokenizer)
    else:
        tokens = text
        
    if urls:
        tokens = remove_urls(tokens)

    if extensions:
        tokens = extract_extensions(tokens)

    if punctuation:
        tokens = remove_punctuation(tokens)
        
    if stopword_list is None:
        stopword_list = []
    
    if stopwords:
        tokens = remove_stopwords(tokens, stopword_list)
        
    if numbers:
        tokens = process_numbers(tokens)

    if stem:
        if stemmer is None:
            stemmer = nltk.stem.SnowballStemmer(language='english')
        tokens = stem_tokens(tokens, stemmer)
        
    return tokens

body_tokens = data['Body'].apply(lambda x : preprocess_text(x, word_tokenizer=word_tokenizer, stopword_list=my_stopwords, stemmer=stemmer))

title_tokens = data['Title'].apply(lambda x : preprocess_text(x, word_tokenizer=word_tokenizer, stopword_list=my_stopwords, stemmer=stemmer))

## Analyse all tokens

all_tokens = []
for tok in body_tokens:
    for t in tok:
        all_tokens.append(t)

len(all_tokens) 

t, c = np.unique(all_tokens, return_counts=True)
unique_tokens = dict(zip(list(t), list(c)))

print("The number of unique tokens: ", len(unique_tokens))

print("Average token count:", np.average(list(unique_tokens.values())))

less_then = 3
rare_tokens = [t for t, c in unique_tokens.items() if c < less_then]
print("Tokens that appear less then", less_then, "times:")
print("====================================")
print("Size: ", len(rare_tokens), "Sample: ", rare_tokens[:50])

more_then = 500
very_frequent_tokens = [(t, c) for t, c in unique_tokens.items() if c > more_then]
print("Tokens that appear more then", more_then, "times:")
print("======================================")
print("Size: ", len(very_frequent_tokens), "Sample: ", very_frequent_tokens[:50])

normal_frequency_tokens = [t for t, c in unique_tokens.items() if c < more_then and c > less_then]
print("Tokens that appear between", less_then, "and", more_then, "times:")
print("==========================================")
print("Size: ", len(normal_frequency_tokens), "Sample: ", normal_frequency_tokens[:50])

### Years in text

def count_occurances(s, tokens):
    count = 0
    for token in tokens:
        if token == s:
            count += 1
    return count

print("Num of occurances of '2008':", count_occurances("2008", all_tokens))
print("Num of occurances of '2010':", count_occurances("2010", all_tokens))

## Compare results with original

index=8

data["Body"][index]

print(body_tokens[index])

data["Title"][index]

print(title_tokens[index])

## Save new data

# Suppress SettingWithCopyWarning
pd.set_option('mode.chained_assignment', None)

data['Body'] = body_tokens
data['Title'] = title_tokens

pd.set_option('mode.chained_assignment', 'warn')

data.head()

file = open(project_path + "data/data_preprocessed.csv", "wb")
pickle.dump(data, file)
file.close()

## Preprocess Tags (for Heuristic)

preprocessed_tags = data['Tags'].apply(lambda x : preprocess_text(x, tokenize=False, html=False, clean=False, urls=False,
                                                                  extensions=False, punctuation=True,  stopwords=False, 
                                                                  numbers=False, stem=True, stemmer=stemmer))

data['Tags'][40]

preprocessed_tags[40]

file = open(project_path + "data/tags_preprocessed.csv", "wb")
pickle.dump(preprocessed_tags, file)
file.close()
