==> 03 Feature selection, Heuristic implemented, Train-Test split stratification problem.ipynb <==
# Feature selection


## Necessary downloads and library imports

from google.colab import drive
drive.mount('/content/drive')

project_path = '/content/drive/My Drive/Colab Notebooks/MATF_ML_project/'

!pip install scikit-multilearn
!pip install ipynb

from ipynb.fs.full.utility import *

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

import pickle

import skmultilearn
from skmultilearn.model_selection import iterative_train_test_split

from scipy import sparse

from sklearn import feature_extraction
from sklearn.model_selection import train_test_split
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import TfidfVectorizer

## Load data

file = open(project_path + "data/data_preprocessed.csv", "rb")
data = pickle.load(file)
file.close()

file = open(project_path + "data/tags_preprocessed.csv", "rb")
tags_preprocessed = pickle.load(file)
file.close()

data.head()

data.shape

X = data[['Title', 'Body']]
y = data['Tags']

tags_preprocessed.head()

## Tag vectorization

def tokenizer_none (x):
    return x
    
def preprocessor_none (x):
    return x

tags_vectorizer = CountVectorizer(analyzer='word', tokenizer=tokenizer_none, 
                                    preprocessor=preprocessor_none, token_pattern=None)
tags_vectorizer.fit(y)

y_vectors = tags_vectorizer.transform(y)

y_vectors

y_vectors.shape

## Train-Test split

X_train, y_train, X_test, y_test = iterative_train_test_split(X.to_numpy(), y_vectors.toarray(), 
                                                              test_size = 0.2)

X_train.shape, y_train.shape, X_test.shape, y_test.shape

## Text Vectorization

For the text vectorization we choose **Tf-Idf** and **Bag-of-Words** vectorizations. 

**How to combine title and body?** The title seems to be an important feature because it is concise and contains the most important information about question. So we will try to exploit it wisely.

---------

We examined two approaches:

1. **Vectorize -> Concatenate** (My idea)

  Vectorize title (0-1 Bag-of-Words), Vectorize body (Tf-Idf) -> Concatenate vectors

  It seems that the frequency is not important, but only the appearance of the words in the title (for example, you probably won't repeat "in c#" many times in question, but this token is very important for predicting the 'c#' tag).
  On the other hand, it is useful to "extract context" from the body because of tags such as 'design' that don't appear explicitly in the text. So we will use Tf-Idf (or some better approach in the future :D)


2. **Concatenate -> Vectorize** (Idea taken from blog <a href="https://medium.datadriveninvestor.com/predicting-tags-for-the-questions-in-stack-overflow-29438367261e">Predicting Tags for the Questions in Stack Overflow</a>)

  Concatenate body and 3 times repeated title -> Vectorize that text

----------


X_train_titles = X_train[:, 0]
X_test_titles = X_test[:, 0]
X_train_bodies = X_train[:, 1]
X_test_bodies = X_test[:, 1]

### 1. Vectorize -> Concatenate

#### Title vectorization

title_vectorizer = CountVectorizer(analyzer='word', binary=True, tokenizer=tokenizer_none, 
                                    preprocessor=preprocessor_none, token_pattern=None, 
                                    min_df=1)
title_vectorizer.fit(X_train_titles)

title_vectors_train = title_vectorizer.transform(X_train_titles)
title_vectors_test = title_vectorizer.transform(X_test_titles)

title_vectors_train.shape, title_vectors_test.shape

print(np.unique(title_vectors_train.toarray(), return_counts=True))
print(np.unique(title_vectors_test.toarray(), return_counts=True))

#### Body vectorization

body_vectorizer = TfidfVectorizer(analyzer='word', tokenizer=tokenizer_none, 
                                    preprocessor=preprocessor_none, token_pattern=None,
                                    min_df=0.0003)
body_vectorizer.fit(X_train_bodies)

body_vectors_train = body_vectorizer.transform(X_train_bodies)
body_vectors_test = body_vectorizer.transform(X_test_bodies)

body_vectors_train.shape, body_vectors_test.shape

#### Concatenate vectors

X_train_concat = sparse.hstack([body_vectors_train, title_vectors_train])
X_test_concat = sparse.hstack([body_vectors_test, title_vectors_test])

X_train_concat.shape, X_test_concat.shape

#### Save data

sparse.save_npz(project_path + "data/X_train_vect_concat.npz", sparse.coo_matrix(X_train_concat))
sparse.save_npz(project_path + "data/y_train_vect_concat.npz", sparse.coo_matrix(y_train))
sparse.save_npz(project_path + "data/X_test_vect_concat.npz", sparse.coo_matrix(X_test_concat))
sparse.save_npz(project_path + "data/y_test_vect_concat.npz", sparse.coo_matrix(y_test))

### 2. Concatenate -> Vectorize

#### Question vectorisation

X_train_question = X_train_bodies + 3 * X_train_titles
X_test_question = X_test_bodies + 3 * X_test_titles

question_vectorizer = TfidfVectorizer(analyzer='word', tokenizer=tokenizer_none, 
                                    preprocessor=preprocessor_none, token_pattern=None,
                                    min_df=0.0003)
question_vectorizer.fit(X_train_question)

question_train_vectors = question_vectorizer.transform(X_train_question)
question_test_vectors = question_vectorizer.transform(X_test_question)

question_train_vectors.shape, question_test_vectors.shape

#### Save data

sparse.save_npz(project_path + "data/X_train_concat_vect.npz", sparse.coo_matrix(question_train_vectors))
sparse.save_npz(project_path + "data/y_train_concat_vect.npz", sparse.coo_matrix(y_train))
sparse.save_npz(project_path + "data/X_test_concat_vect.npz", sparse.coo_matrix(question_test_vectors))
sparse.save_npz(project_path + "data/y_test_concat_vect.npz", sparse.coo_matrix(y_test))

# The Heuristic (Count Tags in text)

Tags such as 'c#', 'java', 'android', file extensions, etc. usually appear in the text. 
So our baseline (heuristic) approach will be counting tag appearances in the text.

tag_counter = CountVectorizer(analyzer='word', tokenizer=tokenizer_none, 
                                    preprocessor=preprocessor_none, token_pattern=None)
tag_counter.fit(tags_preprocessed)

tags_in_question = tag_counter.transform(X['Title'] + X['Body'])

tag_counts = [sum(vect>0) for vect in tags_in_question.toarray()]
print("Average number of tags predicted from a question: %.3f" % (sum(tag_counts)/len(tag_counts)))
print("Max number of tags predicted from a question:", max(tag_counts))

predicated = [[1 if c > 0 else 0 for c in vect] for vect in tags_in_question.toarray()]

print_metrics(y_vectors.toarray(), predicated)

Surprisingly good results!

# Save vectorizers

with open(project_path + 'models/vectorizer_tag.pk', 'wb') as fin:
    pickle.dump(tags_vectorizer, fin)
with open(project_path + 'models/vectorizer_title.pk', 'wb') as fin:
    pickle.dump(title_vectorizer, fin)
with open(project_path + 'models/vectorizer_body.pk', 'wb') as fin:
    pickle.dump(body_vectorizer, fin)
with open(project_path + 'models/vectorizer_question.pk', 'wb') as fin:
    pickle.dump(question_vectorizer, fin)
with open(project_path + 'models/vectorizer_heuristic_tag_counter.pk', 'wb') as fin:
    pickle.dump(tag_counter, fin)

# Experiment: Train-Test Split - Stratification problem

In <a href="http://scikit.ml/stratification.html">Multi-label data stratification - Scikit-Multilearn</a> it is illustrated how `skmultilearn.model_selection.iterative_train_test_split()`, which is adapted to the problem of multi-label classification, preserves the ratio of counts of label combinations in train and test sets. We are curious to illustrate how it preserves the counts of each label.



#### Testing scikit-multilearn method for train-test split (with stratification)

X_train, y_train, X_test, y_test = iterative_train_test_split(X.to_numpy(), y_vectors.toarray(), 
                                                              test_size = 0.5)

tag_counts_train = sum(y_train)
tag_counts_test = sum(y_test)

print(tag_counts_train)

print(tag_counts_test)

print(list(zip(tag_counts_train, tag_counts_test))[0:15])

# differences between counts of tag in train and test sets
diff = [abs(ttrain - ttest) for ttrain, ttest in zip(tag_counts_train, tag_counts_test)]

print("Max difference: ", max(diff))
print("Avg. difference: ", np.average(diff))

print("%.3f %%" % (100 * max(diff) / np.average(tag_counts_train)))

#### Testing scikit-learn method for train-test split with no stratification



X_train, X_test, y_train, y_test = train_test_split(X.to_numpy(), y_vectors.toarray(), 
                                                              test_size = 0.5)

y_train.shape

tag_counts_train = sum(y_train)
tag_counts_test = sum(y_test)

print(tag_counts_train)

print(tag_counts_test)

print(list(zip(tag_counts_train, tag_counts_test))[0:15])

# differences between counts of tag in train and test sets
diff = [abs(ttrain - ttest) for ttrain, ttest in zip(tag_counts_train, tag_counts_test)]

print("Max difference: ", max(diff))
print("Avg. difference: ", np.average(diff))

print("%.3f %%" % (100 * max(diff) / np.average(tag_counts_train)))
