==> 04 Models - Transformation into binary classification.ipynb <==
# Models - Transformation into binary classification

**Transformation techniques used:**

* Binary Relevance
* Classification Chains, with classifier orders:
    * Most frequent first
    * Least frequent first
    * Random order

**Base classification models used:**

* Logistic Regression
* Multinomial Naive Bayes
* Support Vector Machines

## Necessary downloads and library imports

from google.colab import drive
drive.mount('/content/drive')

project_path = '/content/drive/My Drive/Colab Notebooks/MATF_ML_project/'

!pip install scikit-multilearn
!pip install ipynb

from ipynb.fs.full.utility import *

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import random

import pickle
import time
import datetime
import copy
import warnings

from scipy import sparse

import skmultilearn
from skmultilearn.model_selection import iterative_train_test_split
from skmultilearn.problem_transform import BinaryRelevance
from skmultilearn.problem_transform import ClassifierChain

from sklearn.model_selection import GridSearchCV
from sklearn.model_selection import ShuffleSplit
from sklearn.model_selection import PredefinedSplit
from sklearn.linear_model import LogisticRegression
from sklearn.naive_bayes import MultinomialNB
from sklearn.svm import SVC
import sklearn.metrics as metrics
from sklearn.metrics import make_scorer

## Load data

X_train_vect_concat = sparse.load_npz(project_path + 'data/X_train_vect_concat.npz')
y_train_vect_concat = sparse.load_npz(project_path + 'data/y_train_vect_concat.npz')
X_test_vect_concat = sparse.load_npz(project_path + 'data/X_test_vect_concat.npz')
y_test_vect_concat = sparse.load_npz(project_path + 'data/y_test_vect_concat.npz')

X_train_concat_vect = sparse.load_npz(project_path + 'data/X_train_concat_vect.npz')
y_train_concat_vect = sparse.load_npz(project_path + 'data/y_train_concat_vect.npz')
X_test_concat_vect = sparse.load_npz(project_path + 'data/X_test_concat_vect.npz')
y_test_concat_vect = sparse.load_npz(project_path + 'data/y_test_concat_vect.npz')

X_train_vect_concat.shape, y_train_vect_concat.shape, X_test_vect_concat.shape, y_test_vect_concat.shape

X_train_concat_vect.shape, y_train_concat_vect.shape, X_test_concat_vect.shape, y_test_concat_vect.shape

N_labels = y_train_vect_concat.shape[1]
N_labels

data_list = [(X_train_vect_concat, y_train_vect_concat), (X_train_concat_vect, y_train_concat_vect)]

## Binary Relevance

parameters = [
    {
        'classifier': [MultinomialNB()],
    },
    {
        'classifier': [SVC()],
        'classifier__kernel': ['linear'],
        'classifier__verbose': [0],
        'require_dense': [[False, True]],
    },
    {
        'classifier': [LogisticRegression()],
        'classifier__solver': ['lbfgs'],
        'classifier__verbose': [0],
        'require_dense': [[False, True]],
    },
]

When true positive + false positive == 0, precision is undefined. When true positive + false negative == 0, recall is undefined. In such cases, by default the metric will be set to 0, as will f-score, and UndefinedMetricWarning will be raised. We will suppress this warning.

with warnings.catch_warnings():
    warnings.simplefilter("ignore")
    result_binary_relevance = GridSearchCVMultipleFits(BinaryRelevance(), parameters, 
                              cv=ShuffleSplit(test_size=0.20, n_splits=1, random_state=0), 
                              scoring=['f1_weighted', 'f1_micro'],
                              refit='f1_weighted', verbose=0,
                              data_list=data_list)

print(result_binary_relevance)

print("Best F1-score weighted:", result_binary_relevance['best_score_'])

names = ['Binary Relevance - Multinomial Naive Bayes, vectorization -> concatenate\t%.3f',
         'Binary Relevance - Support Vector Machines, vectorization -> concatenate\t%.3f',
         'Binary Relevance - Logistic Regression, vectorization -> concatenate\t\t%.3f',
         'Binary Relevance - Multinomial Naive Bayes, concatenate -> vectorization\t%.3f',
         'Binary Relevance - Support Vector Machines, concatenate -> vectorization\t%.3f',
         'Binary Relevance - Logistic Regression, concatenate -> vectorization\t\t%.3f']

print("--------------------------------------------")
print("F1-scores Micro averaged on validation data:")
print("--------------------------------------------")
for i in range(len(names)):
    print(names[i] % (100 * result_binary_relevance['mean_test_f1_micro'][i]))

print()
print("--------------------------------------")
print("F1-scores Weighted on validation data:")
print("--------------------------------------")
for i in range(len(names)):
    print(names[i] % (100 * result_binary_relevance['mean_test_f1_weighted'][i]))

## Classifier Chains

order_freq_incr = list(range(N_labels))

order_freq_decr = list(range(N_labels-1,-1,-1))

order_random = copy.deepcopy(order_freq_incr)
random.shuffle(order_random)

parameters = [
    {
        'classifier': [SVC()],
        'classifier__kernel': ['linear'],
        'classifier__verbose': [0],
        'order': [order_freq_decr, order_freq_incr, order_random],
        'require_dense': [[False, True]],
    },
    {
        'classifier': [LogisticRegression()],
        'classifier__solver': ['lbfgs'],
        'classifier__verbose': [0],
        'order': [order_freq_decr, order_freq_incr, order_random],
        'require_dense': [[False, True]],
    },
]

When true positive + false positive == 0, precision is undefined. When true positive + false negative == 0, recall is undefined. In such cases, by default the metric will be set to 0, as will f-score, and UndefinedMetricWarning will be raised. We will suppress this warning.

with warnings.catch_warnings():
    warnings.simplefilter("ignore")
    result_classifier_chain = GridSearchCVMultipleFits(ClassifierChain(), parameters, 
                              cv=ShuffleSplit(test_size=0.20, n_splits=1, random_state=0), 
                              scoring=['f1_weighted', 'f1_micro'],
                              refit='f1_weighted', verbose=0,
                              data_list=data_list)

print(result_classifier_chain)

print("Best F1-score weighted:", result_classifier_chain['best_score_'])

names = ['Classifier Chains - Support Vector Machines, vectorization -> concatenate, order: Least frequent first\t%.3f',
         'Classifier Chains - Support Vector Machines, vectorization -> concatenate, order: Most frequent first\t%.3f',
         'Classifier Chains - Support Vector Machines, vectorization -> concatenate, order: Random order\t\t%.3f',
         'Classifier Chains - Logistic Regression, vectorization -> concatenate, order: Least frequent first\t%.3f',
         'Classifier Chains - Logistic Regression, vectorization -> concatenate, order: Most frequent first\t%.3f',
         'Classifier Chains - Logistic Regression, vectorization -> concatenate, order: Random order\t\t%.3f',

         'Classifier Chains - Support Vector Machines, concatenate -> vectorization, order: Least frequent first\t%.3f',
         'Classifier Chains - Support Vector Machines, concatenate -> vectorization, order: Most frequent first\t%.3f',
         'Classifier Chains - Support Vector Machines, concatenate -> vectorization, order: Random order\t\t%.3f',
         'Classifier Chains - Logistic Regression, concatenate -> vectorization, order: Least frequent first\t%.3f',
         'Classifier Chains - Logistic Regression, concatenate -> vectorization, order: Most frequent first\t%.3f',
         'Classifier Chains - Logistic Regression, concatenate -> vectorization, order: Random order\t\t%.3f']

print("--------------------------------------------")
print("F1-scores Micro averaged on validation data:")
print("--------------------------------------------")
for i in range(len(names)):
    print(names[i] % (100 * result_classifier_chain['mean_test_f1_micro'][i]))

print()
print("--------------------------------------")
print("F1-scores Weighted on validation data:")
print("--------------------------------------")
for i in range(len(names)):
    print(names[i] % (100 * result_classifier_chain['mean_test_f1_weighted'][i]))

## Save models

with open(project_path + 'models/model_best_binary_relevance.pk', 'wb') as fin:
    pickle.dump(result_binary_relevance['best_estimator_'], fin)
with open(project_path + 'models/model_best_classifier_chain.pk', 'wb') as fin:
    pickle.dump(result_classifier_chain['best_estimator_'], fin)
